Revert "Fix mismatched definitions for rng_fill_buffer"
This reverts commit 6dbd4707a7fef0edbf991d204df2a3acf931be05.

+++ rng.h
@@ -7,7 +7,7 @@
 
 /**
  * Resets the random number generator and gives it the given seed.
- *
+ * 
  * @param device_testing_context  The device whose RNG should be initialized.
  * @param seed                    The seed value to provide to the RNG.
  */
@@ -27,7 +27,7 @@ void rng_reseed(device_testing_context_type *device_testing_context, unsigned in
  * Obtains a random number from the random number generator.  Since random()
  * only generates 31 bits of random data, this function randomizes the uppermost
  * bit of the result.
- *
+ * 
  * @param device_testing_context  The device whose RNG should be used to
  *                                generate the random number.
  *
@@ -42,11 +42,11 @@ int32_t rng_get_random_number(device_testing_context_type *device_testing_contex
  *                                generate random bytes for the buffer.
  * @param buffer                  A pointer to the buffer to be populated with
  *                                random bytes.
- *
+ * 
  * @param size  The number of bytes to write to the buffer.  Must be a multiple
  *              of 4 -- if it isn't, it will be rounded up to the next multiple
  *              of 4.
 */
-void rng_fill_buffer(device_testing_context_type *device_testing_context, char *buffer, uint64_t size);
+void rng_fill_buffer(device_testing_context_type *device_testing_context, char *buffer, size_t size);
 
 #endif // !defined(RNG_H)
